Split test_butler.py into seven files by subject area
Pure code motion. Every top-level definition is byte-identical to the form
it had before this commit; tests/_migration/verify_motion.py checks that by
concatenating the new files and comparing all of them, and reports no
definition missing, added or changed.

The file keeps its history as test_butler_lifecycle.py, the largest of the
seven.
